+++ src/HolidayCollection.php
@@ -19,12 +19,6 @@ use ArrayIterator;
 use Countable;
 use IteratorAggregate;
 use Niirrty\Date\DateTime;
-use function array_keys;
-use function array_search;
-use function count;
-use function in_array;
-use function is_int;
-use function is_null;
 
 
 /**
@@ -36,34 +30,12 @@ class HolidayCollection implements ArrayAccess, IteratorAggregate, Countable
 
     // <editor-fold desc="= = =   P R I V A T E   F I E L D S   = = = = = = = = = = = = = = = = = = = = = = = = =">
 
-
-    /**
-     * The country name (e.g. 'Deutschland' or 'United Kingdom')
-     *
-     * @type string
-     */
-    private $_countryName;
-
-    /**
-     * The 2 char ISO country ID (e.g.: 'de', 'fr')
-     *
-     * @type string
-     */
-    private $_countryId;
-
     /**
      * All Holiday records.
      *
      * @type Holiday[] Array
      */
-    private $_data;
-
-    /**
-     * The year where the holidays are valid for
-     *
-     * @type int
-     */
-    private $_year;
+    private array $_data;
 
     /**
      * A numeric indicated array that defines the names of all regions for current country.
@@ -72,7 +44,7 @@ class HolidayCollection implements ArrayAccess, IteratorAggregate, Countable
      *
      * @type array
      */
-    private $_regions;
+    private array $_regions;
 
     // </editor-fold>
 
@@ -82,16 +54,14 @@ class HolidayCollection implements ArrayAccess, IteratorAggregate, Countable
     /**
      * HolidayCollection constructor.
      *
-     * @param int    $year        The year where the holidays are valid for
+     * @param int    $year        The year, where the holidays are valid for
      * @param string $countryName The country name (e.g. 'Deutschland' or 'United Kingdom')
      * @param string $countryId   The 2 char ISO country ID (e.g.: 'de', 'fr')
      */
-    public function __construct( int $year, string $countryName, string $countryId )
+    public function __construct( private readonly int $year, private readonly string $countryName,
+                                 private readonly string $countryId )
     {
 
-        $this->_year = $year;
-        $this->_countryName = $countryName;
-        $this->_countryId = $countryId;
         $this->_data = [];
         $this->_regions = [];
 
@@ -111,7 +81,7 @@ class HolidayCollection implements ArrayAccess, IteratorAggregate, Countable
      * @link  http://php.net/manual/en/iteratoraggregate.getiterator.php
      * @return ArrayIterator An instance of an object implementing <b>Iterator</b> or <b>Traversable</b>
      */
-    public function getIterator()
+    public function getIterator() : ArrayIterator
     {
 
         return new ArrayIterator( $this->_data );
@@ -132,7 +102,7 @@ class HolidayCollection implements ArrayAccess, IteratorAggregate, Countable
      *                 was returned.
      * @link  http://php.net/manual/en/arrayaccess.offsetexists.php
      */
-    public function offsetExists( $offset )
+    public function offsetExists( $offset ) : bool
     {
 
         return isset( $this->_data[ $offset ] );
@@ -147,7 +117,7 @@ class HolidayCollection implements ArrayAccess, IteratorAggregate, Countable
      * @return Holiday
      * @link   http://php.net/manual/en/arrayaccess.offsetget.php
      */
-    public function offsetGet( $offset )
+    public function offsetGet( $offset ) : Holiday
     {
 
         return $this->_data[ $offset ];
@@ -163,7 +133,7 @@ class HolidayCollection implements ArrayAccess, IteratorAggregate, Countable
      * @throws Exception
      * @link   http://php.net/manual/en/arrayaccess.offsetset.php
      */
-    public function offsetSet( $offset, $value )
+    public function offsetSet( $offset, $value ) : void
     {
 
         if ( !( $value instanceof Holiday ) )
@@ -171,7 +141,7 @@ class HolidayCollection implements ArrayAccess, IteratorAggregate, Countable
             throw new Exception( 'Can not set an holiday if it is not an Holiday instance!' );
         }
 
-        if ( is_null( $offset ) )
+        if ( \is_null( $offset ) )
         {
             $this->_data[ $value->getIdentifier() ] = $value;
         }
@@ -189,7 +159,7 @@ class HolidayCollection implements ArrayAccess, IteratorAggregate, Countable
      *
      * @link  http://php.net/manual/en/arrayaccess.offsetunset.php
      */
-    public function offsetUnset( $offset )
+    public function offsetUnset( $offset ) : void
     {
 
         unset( $this->_data[ $offset ] );
@@ -207,10 +177,10 @@ class HolidayCollection implements ArrayAccess, IteratorAggregate, Countable
      * @return int The custom count as an integer. The return value is cast to an integer.
      * @link   http://php.net/manual/en/countable.count.php
      */
-    public function count()
+    public function count() : int
     {
 
-        return count( $this->_data );
+        return \count( $this->_data );
 
     }
 
@@ -239,7 +209,7 @@ class HolidayCollection implements ArrayAccess, IteratorAggregate, Countable
     public function getCountryId(): string
     {
 
-        return $this->_countryId;
+        return $this->countryId;
 
     }
 
@@ -251,7 +221,7 @@ class HolidayCollection implements ArrayAccess, IteratorAggregate, Countable
     public function getCountryName(): string
     {
 
-        return $this->_countryName;
+        return $this->countryName;
 
     }
 
@@ -263,7 +233,7 @@ class HolidayCollection implements ArrayAccess, IteratorAggregate, Countable
     public function getYear(): int
     {
 
-        return $this->_year;
+        return $this->year;
 
     }
 
@@ -282,19 +252,19 @@ class HolidayCollection implements ArrayAccess, IteratorAggregate, Countable
     /**
      * Gets if an region with the defined index or name exists.
      *
-     * @param string|int $region The name or index of the required region
+     * @param int|string $region The name or index of the required region
      *
      * @return bool
      */
-    public function hasRegion( $region ): bool
+    public function hasRegion( int|string $region ): bool
     {
 
-        if ( is_int( $region ) )
+        if ( \is_int( $region ) )
         {
             return isset( $this->_regions[ $region ] );
         }
 
-        return ( false !== array_search( $region, $this->_regions ) );
+        return ( \in_array( $region, $this->_regions ) );
 
     }
 
@@ -306,7 +276,7 @@ class HolidayCollection implements ArrayAccess, IteratorAggregate, Countable
     public function hasRegions(): bool
     {
 
-        return count( $this->_regions ) > 0;
+        return \count( $this->_regions ) > 0;
 
     }
 
@@ -334,7 +304,7 @@ class HolidayCollection implements ArrayAccess, IteratorAggregate, Countable
     public function get( string $identifier ): ?Holiday
     {
 
-        if ( !$this->has( $identifier ) )
+        if ( ! $this->has( $identifier ) )
         {
             return null;
         }
@@ -351,7 +321,7 @@ class HolidayCollection implements ArrayAccess, IteratorAggregate, Countable
     public function getIdentifiers(): array
     {
 
-        return array_keys( $this->_data );
+        return \array_keys( $this->_data );
 
     }
 
@@ -430,14 +400,14 @@ class HolidayCollection implements ArrayAccess, IteratorAggregate, Countable
 
         $names = [];
 
-        if ( count( $regionIndexes ) < 1 || $regionIndexes[ 0 ] === -1 )
+        if ( \count( $regionIndexes ) < 1 || $regionIndexes[ 0 ] === -1 )
         {
             return $this->_regions;
         }
 
         foreach ( $regionIndexes as $regionIndex )
         {
-            if ( !isset( $this->_regions[ $regionIndex ] ) )
+            if ( ! isset( $this->_regions[ $regionIndex ] ) )
             {
                 continue;
             }
@@ -455,10 +425,10 @@ class HolidayCollection implements ArrayAccess, IteratorAggregate, Countable
      *
      * @return int|FALSE
      */
-    public function indexOfRegion( string $region )
+    public function indexOfRegion( string $region ) : false|int
     {
 
-        return array_search( $region, $this->_regions );
+        return \array_search( $region, $this->_regions );
 
     }
 
@@ -469,8 +439,9 @@ class HolidayCollection implements ArrayAccess, IteratorAggregate, Countable
      * @param string|null $foundIdentifier Return the holiday identifier if the method return true
      *
      * @return bool
+     * @throws \Throwable
      */
-    public function containsDate( $date, string &$foundIdentifier = null ): bool
+    public function containsDate( mixed $date, string &$foundIdentifier = null ): bool
     {
 
         if ( false === ( $dt = DateTime::Parse( $date ) ) )
@@ -478,7 +449,7 @@ class HolidayCollection implements ArrayAccess, IteratorAggregate, Countable
             return false;
         }
 
-        $dateString = '' . $this->_year . $dt->format( '-m-d' );
+        $dateString = '' . $this->year . $dt->format( '-m-d' );
 
         foreach ( $this->_data as $holiday )
         {
@@ -506,7 +477,7 @@ class HolidayCollection implements ArrayAccess, IteratorAggregate, Countable
     public function containsDay( int $month, int $day, string &$foundIdentifier = null ): bool
     {
 
-        $dt = DateTime::Create( $this->_year, $month, $day );
+        $dt = DateTime::Create( $this->year, $month, $day );
 
         $dateString = $dt->format( 'Y-m-d' );
 
@@ -537,7 +508,7 @@ class HolidayCollection implements ArrayAccess, IteratorAggregate, Countable
     public function startsAt( int $month = 1, int $day = 1 ): HolidayCollection
     {
 
-        $result = ( new HolidayCollection( $this->_year, $this->_countryName, $this->_countryId ) )
+        $result = ( new HolidayCollection( $this->year, $this->countryName, $this->countryId ) )
             ->setRegions( $this->_regions );
 
         if ( $month < 1 )
@@ -549,7 +520,7 @@ class HolidayCollection implements ArrayAccess, IteratorAggregate, Countable
             $day = DateTime::CurrentDay();
         }
 
-        $minDate = DateTime::Create( $this->_year, $month, $day );
+        $minDate = DateTime::Create( $this->year, $month, $day );
 
         foreach ( $this->_data as $identifier => $holiday )
         {
@@ -577,13 +548,13 @@ class HolidayCollection implements ArrayAccess, IteratorAggregate, Countable
     public function byRegionId( int $regionId ): HolidayCollection
     {
 
-        $result = ( new HolidayCollection( $this->_year, $this->_countryName, $this->_countryId ) )
+        $result = ( new HolidayCollection( $this->year, $this->countryName, $this->countryId ) )
             ->setRegions( $this->_regions );
 
         foreach ( $this->_data as $identifier => $holiday )
         {
             $regions = $holiday->getValidRegions();
-            if ( [ -1 ] === $regions || [] === $regions || in_array( $regionId, $regions, true ) )
+            if ( [ -1 ] === $regions || [] === $regions || \in_array( $regionId, $regions, true ) )
             {
                 $result->_data[ $identifier ] = $holiday;
             }
@@ -603,7 +574,7 @@ class HolidayCollection implements ArrayAccess, IteratorAggregate, Countable
     public function byRegionIds( array $regionIds ): HolidayCollection
     {
 
-        $result = ( new HolidayCollection( $this->_year, $this->_countryName, $this->_countryId ) )
+        $result = ( new HolidayCollection( $this->year, $this->countryName, $this->countryId ) )
             ->setRegions( $this->_regions );
 
         foreach ( $this->_data as $identifier => $holiday )
@@ -617,7 +588,7 @@ class HolidayCollection implements ArrayAccess, IteratorAggregate, Countable
             {
                 foreach ( $regionIds as $regionId )
                 {
-                    if ( in_array( $regionId, $regions, true ) )
+                    if ( \in_array( $regionId, $regions, true ) )
                     {
                         $result->_data[ $identifier ] = $holiday;
                     }
@@ -643,7 +614,7 @@ class HolidayCollection implements ArrayAccess, IteratorAggregate, Countable
 
         if ( false === $regionId || 1 > $regionId )
         {
-            return ( new HolidayCollection( $this->_year, $this->_countryName, $this->_countryId ) )
+            return ( new HolidayCollection( $this->year, $this->countryName, $this->countryId ) )
                 ->setRegions( $this->_regions );
         }
 
@@ -682,7 +653,7 @@ class HolidayCollection implements ArrayAccess, IteratorAggregate, Countable
     public function restDays(): HolidayCollection
     {
 
-        $result = ( new HolidayCollection( $this->_year, $this->_countryName, $this->_countryId ) )
+        $result = ( new HolidayCollection( $this->year, $this->countryName, $this->countryId ) )
             ->setRegions( $this->_regions );
 
         foreach ( $this->_data as $identifier => $holiday )
@@ -706,7 +677,7 @@ class HolidayCollection implements ArrayAccess, IteratorAggregate, Countable
     public function noRestDays(): HolidayCollection
     {
 
-        $result = ( new HolidayCollection( $this->_year, $this->_countryName, $this->_countryId ) )
+        $result = ( new HolidayCollection( $this->year, $this->countryName, $this->countryId ) )
             ->setRegions( $this->_regions );
 
         foreach ( $this->_data as $identifier => $holiday )